# openapi/ai-capabilities.yaml
openapi: "3.0.3"
info:
  title: AI 能力 API
  description: |
    AI 能力模块，包含：
    - AI 核心（对话、版本、行为、同步）
    - AI 健康检查
    - AI 监控
    - 插件市场
    - RAG 知识库
    - LLM Provider
    - AI 协作
    - 提示词模板
    - Agent Workflow
    - 高级功能
    - Provider 管理
    - AI 模型
    - 文档解析
  version: "1.0.0"
servers:
  - url: http://localhost:3001
    description: 开发环境
...
paths:
...
  /api/ai/rag/{schemaId}:
    delete:
      tags: [ai-rag]
      summary: 删除 Schema 向量嵌入
      description: 删除指定 Schema 的向量嵌入（不删除源 Schema）
      operationId: deleteRagEmbedding
      parameters:
        - name: schemaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/ai/rag/reindex/{schemaId}:
    post:
      tags: [ai-rag]
      summary: 重建单个 Schema 索引
      operationId: reindexSingleSchema
      parameters:
        - name: schemaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      schemaId:
                        type: string
                      action:
                        type: string
                        enum: ["created", "updated", "skipped"]

  # ==================== LLM Provider ====================
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: object
          properties:
            message:
              type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    workflowKeyAuth:
      type: apiKey
      in: header
      name: X-Workflow-Key